Fixes default empty string conversion to numerical value
Missing attributes will be converted as empty string from EDS Value proto object.
As a result, when a query fetches an optional attribute that could be missing
on some entities, an exception is thrown during conversion.

To avoid throwing an NumerFormatExcpetion, convert empty string to zero
Double / Long data in the Value object, since that would the default value
in proto object as well.