Convert time range to a filter on EntitiesRequest for simplicity.
Currently, the start and end times are first class fields on EntitiesRequest
and the filter for time range is created towards the end in EntityFetcher while
sending the actual query to the data source. However, that means all the logic
that's present in execution tree doesn't consider the time range and hence it
causes some bugs.

This change will keep the execution tree logic as is while also fixing time range
filtering bugs.